backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
chromadb==0.5.23
openai==1.57.0
python-dotenv==1.0.1
numpy<2
